[dt,clkmgr] Move clkmgr dt binding to hw/top/dt instead of ipgen it
In prevision of a future move to DT bindings, only keep a single
copy of the clkmgr DT extension in //hw/top/dt and avoid ipgen'ing
the files which creates multiple copies for no good reason.
